pbs-129--rename: Customer Component 구분
