+++ frontend/src/components/organisms/CustomerQuestionContentItem.js
