+++ frontend/src/components/organisms/CustomerQuestionTypeSelectItem.js
@@ -27,7 +27,7 @@ function QuestionTypeSelectItem() {
                         </label>
                     </div>
                     <div>주문 문의</div>
-
+ 
                     <div className="radio">
                         <label>
                             <input type="radio" value="option2" />
