+++ frontend/src/pages/voc/QuestionAnswer.js
@@ -1,16 +1,16 @@
 import React, { useEffect, useState } from 'react';
 import Header from '../../components/mocules/Header';
 import Path from '../../components/atoms/Path';
-import QuestionReport from '../../components/organisms/QuestionReport';
-import QuestionSearchItem from '../../components/organisms/QuestionSearchItem';
+import CustomerQuestionReport from '../../components/organisms/CustomerQuestionReport';
+import CustomerQuestionSearchItem from '../../components/organisms/CustomerQuestionSearchItem';
+import CustomerQuestionList from '../../components/templates/CustomerQuestionList';
 import Notification from '../../components/mocules/Notification';
 import { Question_Title } from '../../assets/css/Voc.css';
-import QuestionList from '../../components/templates/QuestionList';
 
 import { getAnswerAndQuestionByCustomerId } from '../../apis/api/answer/answer';
 
 function QuestionAnswer() {
-    // 1번 고객으로 테스트
+    // N번 고객으로 테스트
     const [dataList, setDataList] = useState([]);
     const customerId = 1;
     const token = process.env.REACT_APP_JWT_TOKEN;
@@ -38,10 +38,10 @@ function QuestionAnswer() {
         <>
             <Header login={true} inq={false} voc={true} dashboard={false} />
             <Path largeCategory={'VoC'} mediumCategory={'문의 목록'} />
-            <QuestionReport question_total={totalItems} question_ready={readyItems} question_completed={completedItems} />
-            <QuestionSearchItem totalItems={totalItems}/>
+            <CustomerQuestionReport question_total={totalItems} question_ready={readyItems} question_completed={completedItems} />
+            <CustomerQuestionSearchItem totalItems={totalItems}/>
             <div className={Question_Title}>문의 목록</div>
-            <QuestionList dataList={dataList} />
+            <CustomerQuestionList dataList={dataList} />
             <Notification />
         </>
     );
