+++ frontend/src/pages/voc/QuestionRegister.js
@@ -1,16 +1,16 @@
 import React from 'react';
 import Header from '../../components/mocules/Header';
 import Path from '../../components/atoms/Path';
-import QuestionTypeSelectItem from '../../components/organisms/QuestionTypeSelectItem';
-import QuestionContentItem from '../../components/organisms/QuestionContentItem';
+import CustomerQuestionTypeSelectItem from '../../components/organisms/CustomerQuestionTypeSelectItem';
+import CustomerQuestionContentItem from '../../components/organisms/CustomerQuestionContentItem';
 
 function QuestionRegister() {
     return (
         <>
             <Header login={true} inq={false} voc={true} dashboard={false} />
             <Path largeCategory={'VoC'} mediumCategory={'문의하기'} />
-            <QuestionTypeSelectItem />
-            <QuestionContentItem />
+            <CustomerQuestionTypeSelectItem />
+            <CustomerQuestionContentItem />
         </>
     );
 }
